COM-3441: refacto coursecell
