+++ src/core/components/courses/CourseCell/index.vue
@@ -12,54 +12,30 @@
             <q-item-section>{{ info.label }}</q-item-section>
           </q-item>
         </div>
-        <div v-if="course.status === FORTHCOMING" class="additional-infos-container">
-          <q-item class="infos-course-container">
-            <q-item-section :class="['additional-infos', { 'to-plan' : course.slotsToPlan.length }]">
-              <q-icon size="12px" name="mdi-calendar-range" />
-              <q-item-label>{{ formatCourseSlotsInfos }}</q-item-label>
-            </q-item-section>
-          </q-item>
-          <q-item class="infos-course-container">
-            <q-item-section class="additional-infos">
-              <q-icon size="12px" name="mdi-account-multiple" />
-              <q-item-label>({{ traineesCount }})</q-item-label>
-            </q-item-section>
-          </q-item>
-        </div>
+        <forthcoming-section :course="course" />
       </q-card-section>
     </router-link>
-    <q-card-section v-if="course.status === IN_PROGRESS" class="slots-timeline-container">
-      <div :class="['additional-infos', { 'to-plan' : course.slotsToPlan.length }]">
-        {{ slotsHappened }} / {{ course.slots.length + course.slotsToPlan.length }}
-        {{ course.slotsToPlan.length ? ` - ${course.slotsToPlan.length} à planifier` : '' }}
-      </div>
-      <q-item>
-        <q-item-section>
-          <div class="row slots-timeline">
-            <div v-for="(slot, index) in course.slots" :key="index"
-              :class="['col-3', 'slots', { 'slots-happened': happened(slot) }]" />
-            <div v-for="(slot, index) in course.slotsToPlan" :key="course.slots.length + index + 1"
-              class="col-3 slots slots-to-plan" />
-          </div>
-        </q-item-section>
-      </q-item>
-    </q-card-section>
+    <in-progress-section :course="course" />
   </q-card>
 </template>
 
 <script>
-import get from 'lodash/get';
 import { computed, toRefs } from 'vue';
-import { FORTHCOMING, COMPLETED, IN_PROGRESS, SHORT_DURATION_H_MM, TRAINER } from '@data/constants';
+import { FORTHCOMING, COMPLETED, TRAINER } from '@data/constants';
 import { happened, composeCourseName } from '@helpers/courses';
 import { formatQuantity } from '@helpers/utils';
 import CompaniDate from '@helpers/dates/companiDates';
-import CompaniDuration from '@helpers/dates/companiDurations';
-import { getISOTotalDuration, getRoundedDiffInDays } from '@helpers/dates/utils';
+import { getRoundedDiffInDays } from '@helpers/dates/utils';
 import { useCourses } from '@composables/courses';
+import ForthcomingSection from './ForthcomingSection';
+import InProgressSection from './InProgressSection';
 
 export default {
   name: 'CourseDetail',
+  components: {
+    'forthcoming-section': ForthcomingSection,
+    'in-progress-section': InProgressSection,
+  },
   props: {
     course: { type: Object, default: () => ({}) },
   },
@@ -71,29 +47,7 @@ export default {
 
     const courseName = computed(() => composeCourseName(course.value, isVendorInterface));
 
-    const traineesCount = computed(() => get(course.value, 'trainees.length') || 0);
-
-    const slotsHappened = computed(() => course.value.slots.filter(slot => happened(slot)).length);
-
-    const slotsDurationTitle = computed(() => {
-      if (!course.value || !course.value.slots) return '0h';
-
-      const slotsDuration = getISOTotalDuration(course.value.slots.flat());
-
-      return CompaniDuration(slotsDuration).format(SHORT_DURATION_H_MM);
-    });
-
     const courseSlotsCount = computed(() => course.value.slots.length);
-
-    const formatCourseSlotsInfos = computed(() => {
-      const slotsToPlanLength = course.value.slotsToPlan.length;
-      const totalDates = courseSlotsCount.value + slotsToPlanLength;
-
-      return !totalDates
-        ? '0 date'
-        : `${formatQuantity('date', totalDates)},
-          ${slotsToPlanLength ? `dont ${slotsToPlanLength} à planifier, ` : ''}${slotsDurationTitle.value}`;
-    });
 
     const formatNearestDate = computed(() => {
       if (!courseSlotsCount.value && course.value.estimatedStartDate) {
@@ -158,14 +112,8 @@ export default {
     });
 
     return {
-      // Data
-      FORTHCOMING,
-      IN_PROGRESS,
       // Computed
       courseName,
-      traineesCount,
-      slotsHappened,
-      formatCourseSlotsInfos,
       formatNearestDate,
       headerInfo,
       goToBlendedCourseProfile,
@@ -177,12 +125,12 @@ export default {
 </script>
 
 <style lang="sass" scoped>
-.q-card__section
+:deep(.q-card__section)
   height: fit-content
-  &:hover
+  :deep(&:hover)
     cursor: pointer
-  &--vert
-    padding: 10px
+:deep(.q-card__section--vert)
+  padding: 10px
 .title-text
   font-size: 14px
   color: $copper-grey-700
@@ -193,65 +141,20 @@ export default {
   color: $copper-grey-600
 .item-section-container
   display: flex
-.q-item__section--side
+:deep(.q-item__section--side)
   padding: 0px
   margin-right: 5px
-.q-item
+:deep(.q-item)
   padding: 0px
   min-height: auto
   margin-right: 10px
-.infos-course
-  &-nearest-date
-    color: $copper-grey-900 !important
-    font-size: 14px
-  &-container
-    align-items: center
-    margin-top: 10px
-    & > .q-item__section
-      display: flex
-      justify-content: flex-start
-      flex-direction: row
-      &.q-item__section--side
-        margin-right: 10px
-      & > .q-icon
-        margin-right: 5px
-.slots
-  height: 10px
-  flex: 1
-  background-color: $copper-300
-  &-happened
-    background-color: $primary
-  &-to-plan
-    background-color: $secondary
-
-.additional-infos
+.infos-course-nearest-date
+  color: $copper-grey-900 !important
+  font-size: 14px
+:deep(.additional-infos)
   color: $primary
   font-size: 12px
   align-items: flex-end
-  &-container
-    display: flex
-    justify-content: flex-end
-    flex-wrap: wrap
-
-.to-plan
+:deep(.to-plan)
   color: $secondary
-
-.slots-timeline-container
-  display: flex
-  flex-direction: column
-  padding: 0px
-  overflow: hidden
-  .q-item
-    margin-right: 0px
-  .slots-timeline
-    width: 104%
-    margin-left: -2%
-  .slots-timeline > :nth-child(1)
-    margin-left: 0px
-  .slots-timeline > div
-    margin-left: 1px
-  .additional-infos
-    display: flex
-    padding-right: 10px
-    justify-content: flex-end
 </style>
